refactor: Splitting notifier and scheduler for leveragin cronjob
