+++ Dockerfile.notifier
@@ -1,0 +1,15 @@
+FROM vibioh/scratch
+
+ENV ZONEINFO /zoneinfo.zip
+COPY zoneinfo.zip /zoneinfo.zip
+COPY ca-certificates.crt /etc/ssl/certs/ca-certificates.crt
+
+ENTRYPOINT [ "/notifier" ]
+
+ARG VERSION
+ENV VERSION=${VERSION}
+
+ARG TARGETOS
+ARG TARGETARCH
+
+COPY release/notifier_${TARGETOS}_${TARGETARCH} /notifier
